+++ packages/components/components/elvis-datepicker/src/react/elvia-datepicker.tsx
@@ -33,6 +33,8 @@ export interface DatepickerProps {
   isOpen?: boolean;
   className?: string;
   inlineStyle?: { [style: string]: CSSProperties };
+  hasOptionalText?: boolean;
+  showValidation?: boolean;
 }
 
 export const Datepicker: FC<DatepickerProps> = ({
@@ -54,6 +56,8 @@ export const Datepicker: FC<DatepickerProps> = ({
   isOpen = false,
   className,
   inlineStyle,
+  hasOptionalText,
+  showValidation,
 }) => {
   const [selectedDate, setSelectedDate] = useState(value);
   const [initialFocusedDate, setInitialFocusedDate] = useState<Date | null>(null);
@@ -69,9 +73,13 @@ export const Datepicker: FC<DatepickerProps> = ({
   // Used to avoid date-fns from formatting date before date is valid
   const unicodeChar = '­';
 
+  const showError =
+    (showValidation && currErrorMessage !== '') ||
+    (!hasFocus && (customError || (currErrorMessage !== '' && hasHadFocus)));
+
   // Styling
   const datePickerClasses = classnames('ewc-datepicker', {
-    ['ewc-datepicker--error']: !hasFocus && (customError || (currErrorMessage !== '' && hasHadFocus)),
+    ['ewc-datepicker--error']: showError,
     ['ewc-datepicker--compact']: isCompact !== false,
     ['ewc-datepicker--unselected']: value === null,
     ['ewc-datepicker--full-width']: isFullWidth,
@@ -170,9 +178,10 @@ export const Datepicker: FC<DatepickerProps> = ({
     if (customError) {
       return;
     }
-    if (date === null && !isRequired) {
-      setCurrErrorMessage('');
-    } else if (!isValid(date)) {
+
+    setCurrErrorMessage('');
+
+    if (!isValid(date)) {
       if (date === null && isRequired) {
         setCurrErrorMessage('Velg en dato');
       } else if (date !== null) {
@@ -182,8 +191,6 @@ export const Datepicker: FC<DatepickerProps> = ({
       setCurrErrorMessage(`Kan ikke være før ${format(minDate, 'dd.MM.yyyy')}`);
     } else if (date && maxDate && isAfter(date.setHours(0, 0, 0, 0), maxDate.setHours(0, 0, 0, 0))) {
       setCurrErrorMessage(`Kan ikke være etter ${format(maxDate, 'dd.MM.yyyy')}`);
-    } else {
-      setCurrErrorMessage('');
     }
   };
 
@@ -340,7 +347,13 @@ export const Datepicker: FC<DatepickerProps> = ({
       data-testid="datepicker-wrapper"
     >
       {label !== '' && (
-        <label className="ewc-datepicker__label" aria-label={label} data-testid="datepicker-label">
+        <label
+          className={classnames('ewc-datepicker__label', {
+            'e-form-field__label--optional': hasOptionalText && !isRequired,
+          })}
+          aria-label={label}
+          data-testid="datepicker-label"
+        >
           {label}
         </label>
       )}
@@ -390,7 +403,7 @@ export const Datepicker: FC<DatepickerProps> = ({
         </MuiPickersUtilsProvider>
       </ThemeProvider>
 
-      {!hasFocus && (customError || (currErrorMessage !== '' && hasHadFocus)) && (
+      {showError && (
         <div className="ewc-datepicker__error">
           <Icon name="removeCircle" size="xs" color={getColor('red')} inlineStyle={{ marginTop: '4px' }} />
           <div className="ewc-datepicker__helper-text">